requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1

# ── AI ────────────────────────────────────────────────────────────────────────
openai>=1.30.0,<2.0.0

# ── Image ─────────────────────────────────────────────────────────────────────
Pillow>=10.3.0,<12.0.0

# ── Video (optional – app works without it via fallback) ─────────────────────
moviepy==1.0.3

# ── HTTP (Instagram publish module) ──────────────────────────────────────────
httpx>=0.27.0,<1.0.0

# ── SVG logo desteği ──────────────────────────────────────────────────────────
cairosvg>=2.7.0
